Refactor existing email code into a new area.
